World's lousiest DB thread pool.
- We're not even pooling connections yet
- All the logic happens on the worker thread, not just the queries
- We're using custom structs to pass lots of args back and forth

I can't imagine doing the whole app like this.

We could write wrappers to do individual queries, but the problem is that we're using libuv's pool and each worker gets assigned to a thread at random. When we do a whole transaction in one function, that's fine. Otherwise, it's not.