Pulls slightly more tested now. Amazing async_wakeup() implementation.
First, when we spawn a new fiber, use wakeup instead of switch because we need to keep track of the current thread too (in case it isn't the main thread).

Then, wakeup is now optimal. It doesn't use a single timer. It's even more elegant than our reaper is for destroying threads.