Pulling is actually apparently working now, with some caveats.
Like, it's slow as dirt (seems to slow down the more files you add).

Right now we've got it set up to pull from the old version of EarthFS, which uses a slightly different API.

If we want to support real importing from the old version, we'll need to generate meta-files. Normally you'd just get the meta-files from the other repo.