Slight design flaw with pull reader/writer system.
Yes, it's a ring buffer, but each item can be filled in arbitrary order. We have to block until the next entry is filled, but then we can consume as much as possible.

"Reader" and "writer" are slightly confusing because they are from the perspective of the pull, not from the perspective of the ring buffer. The reader reads from the remote repo and writes to the ring buffer, the writer reads from the ring buffer and writes to the database.