Bug of the month. Cached queries were getting used in several places at once.
Everything worked but it caused the wrong parameters to be bound.

And yes, our "just reorder the terms to get a billion-fold speedup" was too good to be true. Now term ordering is maybe a 25% speedup, which makes sense because it's (at best) shaving off one query per iteration out of like four.