Async SQLite wrapper. Just committing to show I tried it, will be reverted.
With the in-process page cache disabled, this is actually very fast:
- Synchronous: 5 seconds
- Async wrapper: 7 seconds
- Async VFS: 25 seconds

However, caching doesn't help it nearly as much:
- Synchronous: 2 seconds
- Async wrapper: 3.5 seconds
- Async VFS: 2 seconds

This shows that async at the DB API layer is the absolute worst place to do it. And this approach won't work for MDB either.

We could still move all of EarthFS into kernel threads and only use async at the HTTP server layer. Then we could use MDB or blocking SQLite. But I'm not sure it's worth it...