Rewrite the filter system in Objective-C.
I'm aware this is going to alienate a lot of people, but it seems like the best way to get the complexity under control for now. Later we can worry about either porting to C++ or converting it back to straight C (either manually or with a ObjC->C compiler).