Don't allow async_pool_enter to fail (or be canceled).
The error handling itself wasn't really a problem, but it caused mixing of DB and UV errors in EFSRepoDBOpen, which quickly became messy. The small benefit of being able to cancel a fiber while it was waiting to enter a thread just wasn't worth it.

In general I suspect that we shouldn't mix DB and UV in low-level functions that report errors. In higher level functions that handle their errors, they're fine.