Copy paste error.
Due to our functions which free and null out the pointer at the same time, this was merely a memory leak rather than a double-free. The logic of this function is a little bit more complicated than I'd like, and I'm not going to fix it because the whole function is deprecated, so overall I'd say not bad.