We still needed another fsync...
Yeah I fucked it up after all. async_fs_link_mkdirp() fsyncs the directories it creates (if any), but it DOESN'T fsync the directory the link is made in.

To make doing that easier, add async_fs_sync_dirname(). We're really nesting our thread pool calls now, which seems to be working well.