Hack to prevent focus from reaching the search button first.
I tried to come up with a way to reorder the two elements without using absolute positioning, but failed. Well, I didn't want to use tables or flexbox or whatever either.

Focusing the button is pretty much never useful since you can submit the form while focused on the search field. Might be a problem for accessibility, not sure.

Also not sure that -1 is widely supported to disable tab focus. Only tested on Firefox.