Submissions known accept a "known URI" for validation.
Surprising it took us this long to figure this out.

We'll eventually accept PUT requests that let the client specify the URI in advance, which both lets us verify as well as potentially is a lot more efficient.

Right now it's only used by pulls. If it fails there, the pull will probably get stuck retrying forever...