Fix massive use-after-free error. Ouch.
SocketPeek() returns a buffer the socket owns. SocketPop() advances through it. If the buffer is "used up," we still can't free it because the caller is still probably using it.

Would the Rust borrow checker have helped here? Yes, if it could've understood what we were actually trying to do. The borrow is non-lexical because the caller is allowed to keep access to the buffer until the next time it peeks. We might've had to use an unsafe block, in which case we'd have had the same problem.