SLNSync (untested)
These 200 lines took a lot of time and effort to write.

The main stumbling block was my own stupidity. I wasn't in the mindset of building up the underlying API to make the sync system possible.

This code isn't tested yet or called from anywhere. The next step is probably integrating it with SLNPull.